focus on what matters
* move redux contract into a central file
* simplify store interface, remove unnecessary interface hierachy and
additional `dispatch` contracts
