+++ src/test/kotlin/org/rekotlin/StoreDispatchTests.kt
@@ -58,74 +58,74 @@ internal class StoreDispatchTests {
         // TODO: testThrowsExceptionWhenReducersDispatch
     }
 
-    /**
-     * it accepts action creators
-     */
-    @Test
-    fun testAcceptsActionCreators() {
-        store.dispatch(SetValueAction(5))
-
-        val doubleValueActionCreator: ActionCreator<TestAppState, StoreType<TestAppState>> = { state, store ->
-            SetValueAction(state.testValue!! * 2)
-        }
-
-        store.dispatch(doubleValueActionCreator)
-        assertEquals(10, store.state.testValue)
-    }
+//    /**
+//     * it accepts action creators
+//     */
+//    @Test
+//    fun testAcceptsActionCreators() {
+//        store.dispatch(SetValueAction(5))
+//
+//        val doubleValueActionCreator: ActionCreator<TestAppState, StoreType<TestAppState>> = { state, store ->
+//            SetValueAction(state.testValue!! * 2)
+//        }
+//
+//        store.dispatch(doubleValueActionCreator)
+//        assertEquals(10, store.state.testValue)
+//    }
 
     /**
      * it accepts async action creators
      */
-    @Test
-    fun testAcceptsAsyncActionCreators() {
-
-        val awaitEntity = CountDownLatch(1)
-
-        val asyncActionCreator: AsyncActionCreator<TestAppState, StoreType<TestAppState>> = { _, _, callback ->
-            thread {
-                // Provide the callback with an action creator
-                callback { _, _ ->
-                    SetValueAction(5)
-                }
-            }
-        }
-
-        val subscriber = CallbackStoreSubscriber<TestAppState> { state ->
-            this.store.state.testValue?.let {
-                assertEquals(5, it)
-                awaitEntity.countDown()
-            }
-        }
-
-        store.subscribe(subscriber)
-        store.dispatch(asyncActionCreator)
-        assertTrue(awaitEntity.await(1, TimeUnit.SECONDS))
-    }
+//    @Test
+//    fun testAcceptsAsyncActionCreators() {
+//
+//        val awaitEntity = CountDownLatch(1)
+//
+//        val asyncActionCreator: AsyncActionCreator<TestAppState, StoreType<TestAppState>> = { _, _, callback ->
+//            thread {
+//                // Provide the callback with an action creator
+//                callback { _, _ ->
+//                    SetValueAction(5)
+//                }
+//            }
+//        }
+//
+//        val subscriber = CallbackStoreSubscriber<TestAppState> { state ->
+//            this.store.state.testValue?.let {
+//                assertEquals(5, it)
+//                awaitEntity.countDown()
+//            }
+//        }
+//
+//        store.subscribe(subscriber)
+//        store.dispatch(asyncActionCreator)
+//        assertTrue(awaitEntity.await(1, TimeUnit.SECONDS))
+//    }
 
     /**
      * it calls the callback once state update from async action is complete
      */
-    @Test
-    fun testCallsCallbackOnce() {
-
-        val awaitEntity = CountDownLatch(1)
-
-        val asyncActionCreator: AsyncActionCreator<TestAppState, StoreType<TestAppState>> = { _, _, callback ->
-            thread {
-                // Provide the callback with an action creator
-                callback { _, _ ->
-                    SetValueAction(5)
-                }
-            }
-        }
-
-        store.dispatch(asyncActionCreator) { newState ->
-            assertEquals(5, this.store.state.testValue)
-            if (newState.testValue == 5) {
-                awaitEntity.countDown()
-            }
-        }
-
-        assertTrue(awaitEntity.await(1, TimeUnit.SECONDS))
-    }
+//    @Test
+//    fun testCallsCallbackOnce() {
+//
+//        val awaitEntity = CountDownLatch(1)
+//
+//        val asyncActionCreator: AsyncActionCreator<TestAppState, StoreType<TestAppState>> = { _, _, callback ->
+//            thread {
+//                // Provide the callback with an action creator
+//                callback { _, _ ->
+//                    SetValueAction(5)
+//                }
+//            }
+//        }
+//
+//        store.dispatch(asyncActionCreator) { newState ->
+//            assertEquals(5, this.store.state.testValue)
+//            if (newState.testValue == 5) {
+//                awaitEntity.countDown()
+//            }
+//        }
+//
+//        assertTrue(awaitEntity.await(1, TimeUnit.SECONDS))
+//    }
 }
